requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14,<3.15
psycopg2-binary>=2.9,<3.0
drf-spectacular[sidecar]>=0.27,<0.28
gunicorn>=21.2,<22.0
python-dotenv>=1.0,<2.0
django-filter
